applications/feature-engineering/requirements.txt
pyspark==3.5.0
redis==5.0.1
boto3==1.34.0
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1
pyyaml==6.0.1
scikit-learn==1.3.2